% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_glimpse.R
\name{glimpse_to_file}
\alias{glimpse_to_file}
\title{Glimpse generation whatever}
\usage{
glimpse_to_file(dataset, output_name = gsub(x = paste0("glimpse_",
  Sys.time(), ".txt"), pattern = " |:|-", replacement = "_"))
}
\arguments{
\item{dataset}{data.frame passed to tibble's glimpse}

\item{output_name}{names of final output file}
}
\value{
No objects are returned
Ouptut will be a text file written to working directory
}
\description{
Glimpse Function implementation from tibble.
Typically you'll receive multiple files before collating them into a single ADS
This function generates glimpse view of the data.frame passed into it
You are strongly advided to pass output_name to function, else it will create text file with time stamp
}
