% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_glimpse.R
\name{glimpse_to_file}
\alias{glimpse_to_file}
\title{Generate glimpse of dataset}
\usage{
glimpse_to_file(dataset, output_filename = "", output_directory = getwd())
}
\arguments{
\item{dataset}{A data.frame}

\item{output_filename}{Name of the output text file (should end in ".txt")
Strongly advised to pass this parameter, else the function's default is "glimpse_timestamp.txt"}

\item{output_directory}{Directory to write glimpse file to. Default is working directory}
}
\description{
Many a times, you will receive multiple datasets to understand and having a glimpse of it stored in test files might come handy while coding.
This function generates glimpse of a data.frame using tibble::glimpse, and write to a text file.
Using same file name for different datasets will append the outputs to a same file.
}
\examples{

glimpse_to_file(dataset = mtcars, output_filename = "glimpse_mtcars.txt",
output_directory = tempdir())
}
